Restructure component organization, clean up unfinished pages
