--- components/SubscribeInputField
@@ -1,2 +1,0 @@
-import React, { useState } from 'react';
-import styled from 'styled-components';
