+++ components/home/IlliniBlockchainMainComponent/StyledDescription.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+const StyledDescription = styled.h3`
+padding-left: 110px;
+margin-top: 0px;
+margin-bottom: 125px;
+color: #ffffff;  
+display: flex;
+justify-content: left;
+font-size: 36px;
+font-weight: 400;
+`
+
+export default StyledDescription
