+++ components/home/IlliniBlockchainMainComponent/StyledTitleOne.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+const StyledTitleOne = styled.title`
+padding-left: 100px;
+padding-bottom: 0px;
+margin-top: 100px;
+margin-bottom: 0px;
+color: #ffffff; 
+display: flex;
+justify-content: left;
+font-size: 138px;
+font-weight: bold;
+`
+
+export default StyledTitleOne
