+++ components/home/IlliniBlockchainMainComponent/StyledTitleTwo.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+const StyledTitleTwo = styled.title`
+padding-left: 100px;
+margin-top: 0px;
+padding-top: 0px;
+color: #ffffff; 
+display: flex;
+justify-content: left;
+font-size: 138px;
+font-weight: bold;
+`
+
+export default StyledTitleTwo
