+++ components/home/IlliniBlockchainMainComponent/index.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import StyledTitleOne from './StyledTitleOne'
+import StyledTitleTwo from './StyledTitleTwo'
+import StyledDescription from './StyledDescription'
+
+const IlliniBlockchainMainComponent = () => (
+  <div>
+    <StyledTitleOne>Illini</StyledTitleOne>
+    <StyledTitleTwo>Blockchain</StyledTitleTwo>
+    <StyledDescription>The premier source for blockchain at UIUC</StyledDescription>
+  </div>
+)
+
+
+export default IlliniBlockchainMainComponent
