+++ components/home/NewsletterBanner/CustomNewsletterBanner.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+const CustomNewsletterBanner = styled.div`
+background-color: #e84a27;
+`
+
+export default CustomNewsletterBanner
