+++ components/home/NewsletterBanner/StyledDescription.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+const StyledDescription = styled.h3`
+padding-left: 110px;
+padding-bottom: 40px;
+margin-top: 0px;
+color: #ffffff; 
+display: flex;
+justify-content: left;
+font-size: 24px;
+font-weight: 400;
+`
+
+export default StyledDescription
