+++ components/home/NewsletterBanner/StyledTitle.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+const StyledTitle = styled.h2`
+padding-left: 110px;
+padding-bottom: 0px;
+padding-top: 25px;
+margin-bottom: 0px;
+color: #ffffff; 
+display: flex;
+justify-content: left;
+font-size: 48px;
+`
+
+export default StyledTitle
