+++ components/home/NewsletterBanner/index.js
@@ -1,31 +1,8 @@
 import React from 'react'
 import styled from 'styled-components'
-
-const StyledTitle = styled.h2`
-padding-left: 110px;
-padding-bottom: 0px;
-padding-top: 25px;
-margin-bottom: 0px;
-color: #ffffff; 
-display: flex;
-justify-content: left;
-font-size: 48px;
-`
-
-const StyledDescription = styled.h3`
-padding-left: 110px;
-padding-bottom: 40px;
-margin-top: 0px;
-color: #ffffff; 
-display: flex;
-justify-content: left;
-font-size: 24px;
-font-weight: 400;
-`
-
-const CustomNewsletterBanner = styled.div`
-background-color: #e84a27;
-`
+import StyledTitle from './StyledTitle'
+import StyledDescription from './StyledDescription'
+import CustomNewsletterBanner from './CustomNewsletterBanner'
 
 // const SubscribeButton = styled.button<ButtonComponentProps>`
 //   padding: 5px 12px;
