+++ components/home/PillarCard/FlatCard.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const FlatCard = styled.div`
+height: 400px;
+width: 32%;
+background-color: #e84a27;
+border-radius: 25px;
+float: left;
+`
+
+export default FlatCard
