+++ components/home/PillarCard/StyledDescription.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const StyledDescription = styled.h4`
+padding-left: 50px;
+padding-right: 50px;
+display: flex;
+justify-content: center;
+font-size: 24px;
+font-weight: lighter;
+`
+
+export default StyledDescription
