+++ components/home/PillarCard/StyledTitle.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const StyledTitle = styled.title`
+padding-top: 25px;
+display: flex;
+justify-content: center;
+align-items: center;
+font-size: 42px;
+font-weight: bold;
+`
+
+export default StyledTitle
