+++ components/home/StyledTitle.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const StyledTitle = styled.h2`
+margin-top: 150px;
+margin-bottom: 100px;
+color: #ffffff; 
+display: flex;
+background-color: #202020;
+justify-content: center;
+align-items: flex;
+font-size: 48px;
+`;
+
+export default StyledTitle;
