+++ components/home/WhatDoWeDoComponent.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const WhatDoWeDoComponent = styled.div`
+margin: auto;
+width: 90%;
+margin-bottom: 150px;
+color: #ffffff; 
+display: flex;
+justify-content: space-between;
+align-items: center;
+`;
+
+export default WhatDoWeDoComponent;
