+++ components/home/index.js
@@ -1,6 +1,5 @@
 export { default as IlliniBlockchainMainComponent } from './IlliniBlockchainMainComponent';
 export { default as NewsletterBanner } from './NewsletterBanner';
-export { default as CommunityCard } from './CommunityCard';
-export { default as EducationCard } from './EducationCard';
-export { default as ProjectsCard } from './ProjectsCard';
 export { default as PillarCard } from './PillarCard';
+export { default as StyledTitle } from './StyledTitle';
+export { default as WhatDoWeDoComponent } from './WhatDoWeDoComponent';
