+++ components/shared-styled/StyledAbout.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const StyledAbout = styled.h3`
+height: 24px;
+background: #202020;
+color: #ffffff; 
+display: flex;
+justify-content: left;
+align-items: center;
+`
+
+export default StyledAbout
