+++ components/shared-styled/StyledHeader.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const StyledHeader = styled.h1`
+height: 138px;
+background: #202020;
+color: #ffffff; 
+display: flex;
+justify-content: center;
+align-items: stretch;
+`
+
+export default StyledHeader
