+++ components/shared-styled/index.js
@@ -1,0 +1,2 @@
+export { default as StyledAbout } from './StyledAbout'
+export { default as StyledHeader } from './StyledHeader'
