+++ pages/index.js
@@ -1,48 +1,11 @@
 import Head from 'next/head'
-import Image from 'next/image'
-import Navbar from '../components/nav'
-import styles from '../styles/Home.module.css'
-import styled from "styled-components"
 import {
-  EducationCard,
-  CommunityCard,
-  ProjectsCard,
   NewsletterBanner,
   IlliniBlockchainMainComponent,
-  PillarCard
+  PillarCard,
+  StyledTitle,
+  WhatDoWeDoComponent
 } from "../components/home";
-
-
-
-const StyledHeader = styled.h2`
-color: #ffffff; 
-display: flex;
-background-color: #202020;
-justify-content: left;
-align-items: start;
-font-size: 138pt;
-`
-
-const StyledTitle = styled.h2`
-margin-top: 150px;
-margin-bottom: 100px;
-color: #ffffff; 
-display: flex;
-background-color: #202020;
-justify-content: center;
-align-items: flex;
-font-size: 48px;
-`
-
-const WhatDoWeDoComponent = styled.div`
-margin: auto;
-width: 90%;
-margin-bottom: 150px;
-color: #ffffff; 
-display: flex;
-justify-content: space-between;
-align-items: center;
-`
 
 export default function Home() {
   return (
@@ -52,9 +15,12 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
+      
       <IlliniBlockchainMainComponent />
+
       <NewsletterBanner />
-      <div><StyledTitle>What do we do?</StyledTitle></div>
+
+      <StyledTitle>What do we do?</StyledTitle>
       <WhatDoWeDoComponent>           
         <PillarCard
           title={'Education'}
